move config to ps1rfid package
