+++ ps1rfid/config_test.go
